up 9/7

+++ SpringBoot/testting-system/src/main/java/com/vti/testtingsystem/service/impl/DepartmentServiceImpl.java
@@ -2,6 +2,8 @@ package com.vti.testtingsystem.service.impl;
 
 import com.vti.testtingsystem.dto.DepartmentDto;
 import com.vti.testtingsystem.entity.Department;
+import com.vti.testtingsystem.form.DepartmentCreateAndUpdateForm;
+import com.vti.testtingsystem.repository.IAccountRepository;
 import com.vti.testtingsystem.repository.IDepartmentRepository;
 import com.vti.testtingsystem.service.IDepartmentService;
 import org.modelmapper.ModelMapper;
@@ -15,6 +17,9 @@ public class DepartmentServiceImpl implements IDepartmentService {
 
     @Autowired
     private IDepartmentRepository departmentRepository;
+
+    @Autowired
+    private IAccountRepository accountRepository;
 
     @Autowired
     private ModelMapper modelMapper;
@@ -32,28 +37,34 @@ public class DepartmentServiceImpl implements IDepartmentService {
     }
 
     @Override
-    public void create(Department department) {
-        //kiểm tra xen tên đã tồn tại chưa
-        if (departmentRepository.existsByName(department.getName())) {
-            throw new RuntimeException("Department tồn tại");
+    public void create(DepartmentCreateAndUpdateForm form) {
+        if (departmentRepository.existsByName(form.getName())) {
+            throw new RuntimeException("Department đã tồn tại ");
         }
+        Department department = new Department();
+        department.setName(form.getName());
         departmentRepository.save(department);
 
     }
 
     @Override
-    public void update(Department department, Integer id) {
-        Department departmentUpdate = departmentRepository.findById(id).orElseThrow(() -> new RuntimeException("Không tìm thấy Department"));
-        if (departmentRepository.existsByNameAndIdNot(department.getName(), id)) {
+    public void update(DepartmentCreateAndUpdateForm form, Integer id) {
+        Department departmentUpdate = departmentRepository.findById(id)
+                .orElseThrow(() -> new RuntimeException("Không tìm thấy Department"));
+        if (departmentRepository.existsByNameAndIdNot(form.getName(), id)) {
             throw new RuntimeException("Department  tồn tại");
         }
-        departmentUpdate.setName(department.getName());
+        departmentUpdate.setName(form.getName());
         departmentRepository.save(departmentUpdate);
     }
 
 
     @Override
     public void deleteByID(Integer id) {
+        departmentRepository.findById(id).orElseThrow(() -> new RuntimeException("Department không tồn tại"));
+        if (accountRepository.existsByDepartmentId(id)) {
+            throw new RuntimeException("Không thể xóa department vì đang có account");
+        }
         departmentRepository.deleteById(id);
     }
